Rust_find/pars/requirements.txt
beautifulsoup4==4.12.2
fake-useragent==1.2.1
psycopg2-binary==2.9.6
requests==2.31.0